feat: pass custom discussion names to single widget component
